Crate boa_tester

Test262 test runner

This crate will run the full ECMAScript test suite (Test262) and report compliance of the boa engine.

Modules

edition ğŸ”’

Edition detection utilities.

exec ğŸ”’

Execution module for the test runner.

read ğŸ”’

Module to read the list of test suites from disk.

results ğŸ”’

Structs

Harness ğŸ”’

All the harness include files.

HarnessFile ğŸ”’

Ignored ğŸ”’

Structure to allow defining ignored tests, features and files that should be ignored even when reading.

Locale ğŸ”’

Locale information structure.

Statistics ğŸ”’

Represents a tests statistic

SuiteResult ğŸ”’

Outcome of a test suite.

Test ğŸ”’

Represents a test.

TestFlags ğŸ”’

TestResult ğŸ”’

Outcome of a test.

TestSuite ğŸ”’

Represents a test suite.

VersionedStats ğŸ”’

Represents tests statistics separated by ECMAScript edition

Enums

Cli ğŸ”’

Boa test262 tester

Outcome ğŸ”’

An outcome for a test.

Phase ğŸ”’

Phase for an error.

TestOutcomeResult ğŸ”’

Functions

main ğŸ”’

Program entry point.

run_test_suite ğŸ”’

Runs the full test suite.