Enum boa_tester::edition::SpecEdition

#[repr(u8)]
pub(crate) enum SpecEdition {
    ES5,
    ES6,
    ES7,
    ES8,
    ES9,
    ES10,
    ES11,
    ES12,
    ES13,
    ES14,
    ESNext,
}

List of ECMAScript editions that can be tested in the test262 repository.

Variants

ES5

ECMAScript 5.1 Edition

https://262.ecma-international.org/5.1

ES6

ECMAScript 6th Edition

https://262.ecma-international.org/6.0

ES7

ECMAScript 7th Edition

https://262.ecma-international.org/7.0

ES8

ECMAScript 8th Edition

https://262.ecma-international.org/8.0

ES9

ECMAScript 9th Edition

https://262.ecma-international.org/9.0

ES10

ECMAScript 10th Edition

https://262.ecma-international.org/10.0

ES11

ECMAScript 11th Edition

https://262.ecma-international.org/11.0

ES12

ECMAScript 12th Edition

https://262.ecma-international.org/12.0

ES13

ECMAScript 13th Edition

https://262.ecma-international.org/13.0

ES14

ECMAScript 14th Edition

https://262.ecma-international.org/14.0

ESNext

The edition being worked on right now.

A draft is currently available here.

Implementations

impl SpecEdition

pub(crate) fn from_test_metadata(metadata: &MetaData) -> Result<Self, Vec<&str>>

Gets the minimum required ECMAScript edition of a test from its metadata.

If the function finds unknown features in metadata, returns an Err(Vec<&str>) containing the list of unknown features.

pub(crate) fn all_editions() -> impl Iterator<Item = Self>

Gets an iterator of all currently available editions.