Function boa_tester::exec::js262::gc

fn gc(
    _this: &JsValue,
    _: &[JsValue],
    _context: &mut Context<'_>
) -> JsResult<JsValue>

The $262.gc() function.

Wraps the host’s garbage collection invocation mechanism, if such a capability exists. Must throw an exception if no capability exists. This is necessary for testing the semantics of any feature that relies on garbage collection, e.g. the WeakRef API.