Crate boa_tester

Test262 test runner

This crate will run the full ECMAScript test suite (Test262) and report compliance of the boa context.

Modules

exec
Execution module for the test runner.

read
Module to read the list of test suites from disk.

results

Structs

Harness
All the harness include files.

Ignored
Structure to allow defining ignored tests, features and files that should be ignored even when reading.

Locale
Locale information structure.

SuiteResult
Outcome of a test suite.

Test
Represents a test.

TestFlags

TestResult
Outcome of a test.

TestSuite
Represents a test suite.

Enums

Cli
Boa test262 tester

Outcome
An outcome for a test.

Phase
Phase for an error.

TestOutcomeResult

Statics

IGNORED
List of ignored tests.

Functions

main
Program entry point.

run_test_suite
Runs the full test suite.