Crate boa_tester

Test262 test runner

This crate will run the full ECMAScript test suite (Test262) and report compliance of the boa context.

Modules

exec ğŸ”’

Execution module for the test runner.

read ğŸ”’

Module to read the list of test suites from disk.

results ğŸ”’

Structs

Harness ğŸ”’

All the harness include files.

Ignored ğŸ”’

Structure to allow defining ignored tests, features and files that should be ignored even when reading.

Locale ğŸ”’

Locale information structure.

SuiteResult ğŸ”’

Outcome of a test suite.

Test ğŸ”’

Represents a test.

TestFlags ğŸ”’

TestResult ğŸ”’

Outcome of a test.

TestSuite ğŸ”’

Represents a test suite.

Enums

Cli ğŸ”’

Boa test262 tester

Outcome ğŸ”’

An outcome for a test.

Phase ğŸ”’

Phase for an error.

TestOutcomeResult ğŸ”’

Statics

IGNORED ğŸ”’

List of ignored tests.

Functions

main ğŸ”’

Program entry point.

run_test_suite ğŸ”’

Runs the full test suite.