Module boa_tester::results

Structs

ReducedResultInfo
Structure to store full result information.

ResultInfo
Structure to store full result information.

Constants

LATEST_FILE_NAME
File name of the "latest results" JSON file.

RESULTS_FILE_NAME
File name of the "all results" JSON file.

Functions

compare_results
Compares the results of two test suite runs.

get_test262_commit
Gets the commit OID of the test262 submodule.

update_gh_pages_repo
Updates the GitHub pages repository by pulling latest changes before writing the new things.

write_json
Writes the results of running the test suite to the given JSON output file.