Struct boa_tester::Ignored

pub(crate) struct Ignored {
    pub(crate) tests: FxHashSet<Box<str>>,
    pub(crate) features: FxHashSet<Box<str>>,
    pub(crate) flags: TestFlags,
}

Structure to allow defining ignored tests, features and files that should be ignored even when reading.

Fields
tests: FxHashSet<Box<str>>features: FxHashSet<Box<str>>flags: TestFlags
Implementations

impl Ignored

pub(crate) fn contains_test(&self, test: &str) -> bool

Checks if the ignore list contains the given test name in the list of tests to ignore.

pub(crate) fn contains_feature(&self, feature: &str) -> bool

Checks if the ignore list contains the given feature name in the list of features to ignore.

pub(crate) const fn contains_any_flag(&self, flags: TestFlags) -> bool